Specify u64 for parse

+++ ch3/s3.2/rust-types/src/main.rs
@@ -1,4 +1,4 @@
 fn main() {
-    let guess: u32 = "42".parse().expect("Not a number!");
+    let guess: u64 = "42".parse().expect("Not a number!");
     println!("guess is {guess}");
 }
